+++ WoWsShipBuilder.Core/DataProvider/CultureDetails.cs
@@ -1,0 +1,7 @@
+﻿using System;
+using System.Globalization;
+
+namespace WoWsShipBuilder.Core.DataProvider
+{
+    public sealed record CultureDetails(CultureInfo CultureInfo, string LocalizationFileName);
+}