+++ WoWsShipBuilder.UI/UserControls/MessageBox.axaml.cs
@@ -40,13 +40,14 @@ namespace WoWsShipBuilder.UI.UserControls
             Warning,
         }
 
-        public static Task<MessageBoxResult> Show(Window? parent, string text, string title, MessageBoxButtons buttons, MessageBoxIcon icon = MessageBoxIcon.None, int width = 300, int height = 150)
+        public static Task<MessageBoxResult> Show(Window? parent, string text, string title, MessageBoxButtons buttons, MessageBoxIcon icon = MessageBoxIcon.None, int width = 300, int height = 150, SizeToContent sizeToContent = SizeToContent.Manual)
         {
             var msgbox = new MessageBox
             {
                 ShowActivated = true,
                 ShowInTaskbar = true,
                 Title = title,
+                SizeToContent = sizeToContent,
             };
 
             msgbox.FindControl<AppHeader>("Header").Title = title;